fixbug: fix the compile time error
After upgrading the dlog module, neard faced the compile time error
since LOG_ON() macro basically does not return anything. However, some
macros such as NEARDAL_TRACE_ERR() usg this macro to check whether dlog
is enable or not. This is definitely a bug so this patch fixes this issue.

Change-Id: I3d5e61fc245b4160b323c41b9d4ec71cab0fe3fb
Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>